Implement TimelineLogicService for clip management
- Introduced `TimelineLogicService` to encapsulate clip placement and timeline calculations, enhancing separation of concerns.
- Updated `TimelineViewModel` to utilize `TimelineLogicService` for clip placement and overlap management.
- Refactored commands (`AddClipCommand`, `MoveClipCommand`, `ResizeClipCommand`, `TrimOverlapCommand`) to leverage the new service for improved clarity and maintainability.
- Removed redundant methods from `TimelineViewModel` related to clip calculations, streamlining the codebase.
