+++ lib/viewmodels/timeline_viewmodel.dart
@@ -12,6 +12,7 @@ import 'package:watch_it/watch_it.dart';
 import 'commands/timeline_command.dart';
 import 'commands/add_clip_command.dart';
 import 'package:flipedit/services/undo_redo_service.dart';
+import 'package:flipedit/services/timeline_logic_service.dart';
 
 class TimelineViewModel {
   // Add a tag for logging within this class
@@ -20,6 +21,7 @@ class TimelineViewModel {
   final ProjectDatabaseService _projectDatabaseService =
       di<ProjectDatabaseService>();
   final UndoRedoService _undoRedoService = di<UndoRedoService>();
+  final TimelineLogicService _timelineLogicService = di<TimelineLogicService>();
 
   final ValueNotifier<List<ClipModel>> clipsNotifier =
       ValueNotifier<List<ClipModel>>([]);
@@ -58,8 +60,6 @@ class TimelineViewModel {
     EditMode.select,
   );
 
-  Timer? _frameUpdateTimer;
-
   // Helper to set edit mode and notify
   void setEditMode(EditMode mode) {
     if (currentEditMode.value != mode) {
@@ -67,12 +67,9 @@ class TimelineViewModel {
     }
   }
 
-  Timer? _playbackTimer;
   StreamSubscription? _controllerPositionSubscription;
 
   StreamSubscription? _clipStreamSubscription;
-
-  late final VoidCallback _debouncedFrameUpdate;
 
   /// Executes a timeline command and refreshes undo stack
   Future<void> runCommand(TimelineCommand cmd) async {
@@ -103,9 +100,6 @@ class TimelineViewModel {
 
     isPlayingNotifier.value = true;
     logger.logInfo('▶️ Starting playback from frame $currentFrame', _logTag);
-
-    // Start timer for playback
-    _startPlaybackTimer();
   }
 
   // Removed method to update current frame based on engine position during playback
@@ -117,8 +111,6 @@ class TimelineViewModel {
 
     isPlayingNotifier.value = false;
     logger.logInfo('⏹️ Stopping playback at frame $currentFrame', _logTag);
-
-    _stopPlaybackTimer();
   }
 
   Future<void> loadClipsForProject(int projectId) async {
@@ -147,209 +139,6 @@ class TimelineViewModel {
     }
 
     await refreshClips();
-  }
-
-  /// Calculates exact frame position from pixel coordinates on the timeline
-  int calculateFramePosition(
-    double pixelPosition,
-    double scrollOffset,
-    double zoom,
-  ) {
-    final adjustedPosition = pixelPosition + scrollOffset;
-    final frameWidth = 5.0 * zoom; // 5px per frame at 1.0 zoom
-
-    final framePosition = (adjustedPosition / frameWidth).floor();
-    return framePosition < 0 ? 0 : framePosition;
-  }
-
-  /// Converts a frame position to milliseconds (based on standard 30fps)
-  int frameToMs(int framePosition) {
-    return ClipModel.framesToMs(framePosition);
-  }
-
-  /// Calculates millisecond position directly from pixel coordinates
-  int calculateMsPositionFromPixels(
-    double pixelPosition,
-    double scrollOffset,
-    double zoom,
-  ) {
-    final framePosition = calculateFramePosition(
-      pixelPosition,
-      scrollOffset,
-      zoom,
-    );
-    return frameToMs(framePosition);
-  }
-
-  /// Calculates placement for a clip on a track, handling overlaps with neighbors
-  /// Returns placement information without performing database operations
-  Map<String, dynamic> prepareClipPlacement({
-    int? clipId, // If updating an existing clip
-    required int trackId,
-    required ClipType type,
-    required String sourcePath,
-    required int startTimeOnTrackMs,
-    required int startTimeInSourceMs,
-    required int endTimeInSourceMs,
-  }) {
-    final newClipDuration = endTimeInSourceMs - startTimeInSourceMs;
-    int newStart = startTimeOnTrackMs;
-    int newEnd = startTimeOnTrackMs + newClipDuration;
-    
-    // 1. Gather and sort neighbors
-    final neighbors =
-        clips
-            .where(
-              (c) =>
-                  c.trackId == trackId &&
-                  (clipId == null || c.databaseId != clipId),
-            )
-            .toList()
-          ..sort(
-            (a, b) => a.startTimeOnTrackMs.compareTo(b.startTimeOnTrackMs),
-          );
-          
-    // 2. Prepare neighbor modifications (no database operations)
-    List<ClipModel> updatedClips = List<ClipModel>.from(clips);
-    List<Map<String, dynamic>> clipUpdates = [];
-    List<int> clipsToRemove = [];
-    
-    for (final neighbor in neighbors) {
-      final ns = neighbor.startTimeOnTrackMs;
-      final ne = neighbor.startTimeOnTrackMs + neighbor.durationMs;
-      
-      if (ne <= newStart || ns >= newEnd) continue; // No overlap
-      
-      if (ns >= newStart && ne <= newEnd) {
-        // Fully covered: mark for removal
-        updatedClips.removeWhere((c) => c.databaseId == neighbor.databaseId);
-        clipsToRemove.add(neighbor.databaseId!);
-      } else if (ns < newStart && ne > newStart && ne <= newEnd) {
-        // Overlap on right: trim neighbor's end to the intersection
-        final updated = neighbor.copyWith(
-          endTimeInSourceMs: neighbor.startTimeInSourceMs + (newStart - ns),
-        );
-        updatedClips[updatedClips.indexWhere(
-              (c) => c.databaseId == neighbor.databaseId,
-            )] = updated;
-        
-        clipUpdates.add({
-          'id': neighbor.databaseId!,
-          'fields': {'endTimeInSourceMs': neighbor.startTimeInSourceMs + (newStart - ns)},
-        });
-      } else if (ns >= newStart && ns < newEnd && ne > newEnd) {
-        // Overlap on left: trim neighbor's start to the intersection
-        final updated = neighbor.copyWith(
-          startTimeInSourceMs: neighbor.startTimeInSourceMs + (newEnd - ns),
-          startTimeOnTrackMs: newEnd,
-        );
-        updatedClips[updatedClips.indexWhere(
-              (c) => c.databaseId == neighbor.databaseId,
-            )] = updated;
-            
-        clipUpdates.add({
-          'id': neighbor.databaseId!,
-          'fields': {
-            'startTimeInSourceMs': neighbor.startTimeInSourceMs + (newEnd - ns),
-            'startTimeOnTrackMs': newEnd,
-          },
-        });
-      } else if (ns < newStart && ne > newEnd) {
-        // Moved clip is fully inside neighbor: trim neighbor's end to newStart (left part remains)
-        final updated = neighbor.copyWith(
-          endTimeInSourceMs: neighbor.startTimeInSourceMs + (newStart - ns),
-        );
-        updatedClips[updatedClips.indexWhere(
-              (c) => c.databaseId == neighbor.databaseId,
-            )] = updated;
-            
-        clipUpdates.add({
-          'id': neighbor.databaseId!,
-          'fields': {'endTimeInSourceMs': neighbor.startTimeInSourceMs + (newStart - ns)},
-        });
-      }
-      // NEW CASE: If the left neighbor's end overlaps the new start, trim its end to newStart
-      else if (ne > newStart && ne <= newEnd && ns < newStart) {
-        final updated = neighbor.copyWith(
-          endTimeInSourceMs: neighbor.startTimeInSourceMs + (newStart - ns),
-        );
-        updatedClips[updatedClips.indexWhere(
-              (c) => c.databaseId == neighbor.databaseId,
-            )] = updated;
-            
-        clipUpdates.add({
-          'id': neighbor.databaseId!,
-          'fields': {'endTimeInSourceMs': neighbor.startTimeInSourceMs + (newStart - ns)},
-        });
-      }
-    }
-    
-    // 3. Clamp new clip to available space
-    int clampLeft = 0;
-    int clampRight = 1 << 30;
-    for (final neighbor in neighbors) {
-      final ns = neighbor.startTimeOnTrackMs;
-      final ne = neighbor.startTimeOnTrackMs + neighbor.durationMs;
-      if (ne <= newStart) {
-        if (ne > clampLeft) clampLeft = ne;
-      }
-      if (ns >= newEnd) {
-        if (ns < clampRight) clampRight = ns;
-      }
-    }
-    newStart = newStart.clamp(clampLeft, clampRight - 1);
-    newEnd = (newStart + newClipDuration).clamp(clampLeft + 1, clampRight);
-    
-    if (newEnd <= newStart) {
-      return {'success': false};
-    }
-    
-    // 4. Prepare new clip data
-    Map<String, dynamic> newClipData = {
-      'trackId': trackId,
-      'type': type,
-      'sourcePath': sourcePath,
-      'startTimeOnTrackMs': newStart,
-      'startTimeInSourceMs': startTimeInSourceMs,
-      'endTimeInSourceMs': startTimeInSourceMs + (newEnd - newStart),
-    };
-    
-    // For updating existing clip
-    if (clipId != null) {
-      final idx = updatedClips.indexWhere((c) => c.databaseId == clipId);
-      if (idx != -1) {
-        updatedClips[idx] = updatedClips[idx].copyWith(
-          trackId: trackId,
-          startTimeOnTrackMs: newStart,
-          startTimeInSourceMs: startTimeInSourceMs,
-          endTimeInSourceMs: startTimeInSourceMs + (newEnd - newStart),
-        );
-      }
-    } else {
-      // For new clip, prepare model for optimistic UI update
-      ClipModel newClipModel = ClipModel(
-        databaseId: -1, // Temporary ID, will be replaced with actual DB ID
-        trackId: trackId,
-        name: '',
-        type: type,
-        sourcePath: sourcePath,
-        startTimeInSourceMs: startTimeInSourceMs,
-        endTimeInSourceMs: startTimeInSourceMs + (newEnd - newStart),
-        startTimeOnTrackMs: newStart,
-        effects: [],
-        metadata: {},
-      );
-      updatedClips.add(newClipModel);
-    }
-    
-    return {
-      'success': true,
-      'newClipData': newClipData,
-      'clipId': clipId,
-      'updatedClips': updatedClips,
-      'clipUpdates': clipUpdates,
-      'clipsToRemove': clipsToRemove,
-    };
   }
   
   /// Updates the UI state after clip placement (called by commands after persistence)
@@ -404,7 +193,8 @@ class TimelineViewModel {
     } else {
       // For existing clips, use the old direct approach for now
       // This could be refactored to use a MoveClipCommand or similar
-      final placement = prepareClipPlacement(
+      final placement = _timelineLogicService.prepareClipPlacement(
+        clips: clips, // Pass the current clips
         clipId: clipId,
         trackId: trackId,
         type: type,
@@ -544,26 +334,6 @@ class TimelineViewModel {
     recalculateAndUpdateTotalFrames(); // Updated call
   }
 
-  void _startPlaybackTimer() {
-    _stopPlaybackTimer();
-    if (isPlayingNotifier.value) {
-      _playbackTimer = Timer.periodic(
-        Duration(milliseconds: (1000 / kDefaultFrameRate).round()),
-        (_) {
-          // Removed _updateFrameFromEngine() as playhead is now fixed
-          _debouncedFrameUpdate();
-        },
-      );
-    }
-  }
-
-  void _stopPlaybackTimer() {
-    _playbackTimer?.cancel();
-    _playbackTimer = null;
-    _frameUpdateTimer?.cancel();
-    _frameUpdateTimer = null;
-  }
-
   int _calculateTotalFrames() {
     if (clipsNotifier.value.isEmpty) {
       return 0;
@@ -610,184 +380,8 @@ class TimelineViewModel {
 
     trackContentHorizontalScrollController.dispose();
 
-    _stopPlaybackTimer();
     _controllerPositionSubscription?.cancel();
     _clipStreamSubscription?.cancel();
   }
-
-  /// Trims, removes, or splits clips that overlap with [startMs, endMs) on [trackId]. Optionally excludes a clip by ID.
-  Future<void> trimOrRemoveOverlappingClips(
-    int trackId,
-    int startMs,
-    int endMs, [
-    int? excludeClipId,
-  ]) async {
-    final overlapping = getOverlappingClips(
-      trackId,
-      startMs,
-      endMs,
-      excludeClipId,
-    );
-    for (final clip in overlapping) {
-      final clipStart = clip.startTimeOnTrackMs;
-      final clipEnd = clip.startTimeOnTrackMs + clip.durationMs;
-      // Fully covered: remove
-      if (clipStart >= startMs && clipEnd <= endMs) {
-        // Use direct DAO call + recalculate for internal utility method
-        await _projectDatabaseService.clipDao!.deleteClip(clip.databaseId!);
-        recalculateAndUpdateTotalFrames(); // Ensure total frames are updated
-      } else if (clipStart < endMs && clipEnd > endMs) {
-        // Overlap on left: trim neighbor's end (neighbor is to the right of the new clip)
-        if (clip.databaseId != null) {
-          final neighborClip = clip; // Use a clearer variable name
-          final amountToTrimMs =
-              endMs -
-              neighborClip
-                  .startTimeOnTrackMs; // Calculate trim amount explicitly
-
-          // Ensure we don't trim more than the clip's duration
-          if (amountToTrimMs >= neighborClip.durationMs) {
-            // If the overlap implies the entire neighbor clip should be removed
-            await _projectDatabaseService.clipDao!.deleteClip(
-              neighborClip.databaseId!,
-            );
-            // Log this removal for clarity
-            logger.logInfo(
-              'Neighbor clip ${neighborClip.databaseId} fully overlapped and removed.',
-              _logTag,
-            );
-          } else if (amountToTrimMs > 0) {
-            // Only update if there's actually something to trim
-            final newStartTimeInSourceMs =
-                neighborClip.startTimeInSourceMs + amountToTrimMs;
-            final newStartTimeOnTrackMs =
-                endMs; // Set the neighbor's start time to the new clip's end time
-
-            await _projectDatabaseService.clipDao!.updateClipFields(
-              neighborClip.databaseId!,
-              {
-                'startTimeInSourceMs': newStartTimeInSourceMs,
-                'startTimeOnTrackMs':
-                    newStartTimeOnTrackMs, // Update track start time
-              },
-              log: false,
-            );
-            // Log the update
-            logger.logDebug(
-              'Neighbor clip ${neighborClip.databaseId} trimmed (left): new track start $newStartTimeOnTrackMs ms, new source start $newStartTimeInSourceMs ms.',
-              _logTag,
-            );
-          } else {
-            // Log if no trim was needed (e.g., endMs exactly matched clipStart)
-            logger.logDebug(
-              'Neighbor clip ${neighborClip.databaseId} touches new clip end, no trim needed.',
-              _logTag,
-            );
-          }
-        }
-      } else if (clipStart < startMs && clipEnd > startMs) {
-        // Overlap on right: trim neighbor's end
-        await _projectDatabaseService.clipDao!.updateClipFields(
-          clip.databaseId!,
-          {
-            'endTimeInSourceMs':
-                clip.startTimeInSourceMs + (startMs - clipStart),
-          },
-          log: false,
-        );
-      } else if (clipStart < endMs && clipEnd > endMs) {
-        // Overlap on left: trim neighbor's start
-        await _projectDatabaseService.clipDao!
-            .updateClipFields(clip.databaseId!, {
-              'startTimeInSourceMs':
-                  clip.startTimeInSourceMs + (endMs - clipStart),
-              'startTimeOnTrackMs': endMs,
-            }, log: false);
-      }
-    }
-    // await refreshClips(); // Removed immediate refresh - rely on stream
-  }
-
-  /// Returns all clips on the same track that overlap with [startMs, endMs). Optionally excludes a clip by ID.
-  List<ClipModel> getOverlappingClips(
-    int trackId,
-    int startMs,
-    int endMs, [
-    int? excludeClipId,
-  ]) {
-    return clips.where((clip) {
-      if (clip.trackId != trackId) return false;
-      if (excludeClipId != null && clip.databaseId == excludeClipId)
-        return false;
-      final clipStart = clip.startTimeOnTrackMs;
-      final clipEnd = clip.startTimeOnTrackMs + clip.durationMs;
-      // Overlap if ranges intersect
-      return clipStart < endMs && clipEnd > startMs;
-    }).toList();
-  }
-
-  /// Returns a preview of the timeline clips as if a clip were dragged to a new position, applying trimming logic in-memory only.
-  List<ClipModel> getPreviewClipsForDrag({
-    required int clipId,
-    required int targetTrackId,
-    required int targetStartTimeOnTrackMs,
-  }) {
-    final original = clips;
-    final dragged = original.firstWhere((c) => c.databaseId == clipId);
-    final newClipDuration = dragged.durationMs;
-    int newStart = targetStartTimeOnTrackMs;
-    int newEnd = targetStartTimeOnTrackMs + newClipDuration;
-    // Remove the dragged clip from the list
-    final others = original.where((c) => c.databaseId != clipId).toList();
-    List<ClipModel> preview = [];
-    for (final neighbor in others) {
-      if (neighbor.trackId != targetTrackId) {
-        preview.add(neighbor);
-        continue;
-      }
-      final ns = neighbor.startTimeOnTrackMs;
-      final ne = neighbor.startTimeOnTrackMs + neighbor.durationMs;
-      if (ne <= newStart || ns >= newEnd) {
-        preview.add(neighbor);
-      } else if (ns >= newStart && ne <= newEnd) {
-        // Fully covered: remove
-        continue;
-      } else if (ns < newStart && ne > newStart && ne <= newEnd) {
-        // Overlap on right: trim neighbor's end
-        preview.add(
-          neighbor.copyWith(
-            endTimeInSourceMs: neighbor.startTimeInSourceMs + (newStart - ns),
-          ),
-        );
-      } else if (ns >= newStart && ns < newEnd && ne > newEnd) {
-        // Overlap on left: trim neighbor's start
-        preview.add(
-          neighbor.copyWith(
-            startTimeInSourceMs: neighbor.startTimeInSourceMs + (newEnd - ns),
-            startTimeOnTrackMs: newEnd,
-          ),
-        );
-      } else if (ns < newStart && ne > newEnd) {
-        // Dragged clip is fully inside neighbor: only left part remains (trim at newStart)
-        preview.add(
-          neighbor.copyWith(
-            endTimeInSourceMs: neighbor.startTimeInSourceMs + (newStart - ns),
-          ),
-        );
-      }
-    }
-    // Add the dragged clip at the preview position
-    preview.add(
-      dragged.copyWith(
-        trackId: targetTrackId,
-        startTimeOnTrackMs: newStart,
-        // Optionally update startTimeInSourceMs/endTimeInSourceMs if you want to preview source trim
-      ),
-    );
-    // Sort by start time
-    preview.sort(
-      (a, b) => a.startTimeOnTrackMs.compareTo(b.startTimeOnTrackMs),
-    );
-    return preview;
-  }
 }
+
